requirements.txt
lightning==2.2.5
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.4
pandas==2.2.2
pytorch-forecasting==1.0.0
pytorch-lightning==2.2.5
pytorch_optimizer==2.12.0
requests==2.32.3
requests-cache==1.2.0
retry-requests==2.0.0
schedule==1.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torchmetrics==1.4.0.post0
dvc==3.51.2
dvc-gdrive==3.0.1
hydra-core==1.3.2
pre-commit==3.7.1
